Add Elastic example  (#12)
* wip: add Elastic search

* wip: es error Unimplemented type: TimestampType

* fix: timestamp should be Long type

* chores: logs & config

* readme: update for elastic
commited in JFK airport

+++ src/main/scala/radio/Song.scala
@@ -2,7 +2,7 @@ package radio
 
 import java.sql.Timestamp
 
-case class Song(timestamp:Int, humanDate:Long, year:Int, month:Int, day:Int, hour:Int, minute: Int, artist:String, allArtists: String, title:String, radio:String)
+case class Song(timestamp: Long, humanDate:Long, year:Int, month:Int, day:Int, hour:Int, minute: Int, artist:String, allArtists: String, title:String, radio:String)
 
 case class SimpleSong(title: String, artist: String, radio: String)
